Create new dinamic components Multimedia in Home view
