+++ src/Componentes/ElementosDinamicos/CarrouselMultimediaProblematica.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const CarrouselMultimediaProblematica = () => {
+  return (
+    <div>
+      Aqui va el Carrousel
+    </div>
+  )
+}
+
+export default CarrouselMultimediaProblematica;
